clean: minor refactor
